\name{packagecasnit-package}
\alias{packagecasnit-package}
\alias{packagecasnit}
\docType{package}
\title{
\packageTitle{packagecasnit}
}
\description{
\packageDescription{packagecasnit}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{packagecasnit}
\packageIndices{packagecasnit}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{packagecasnit}

Maintainer: \packageMaintainer{packagecasnit}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
